requirements.txt
certifi==2017.4.17
chardet==3.0.4
future==0.16.0
idna==2.5
itchat==1.2.32
olefile==0.44
Pillow==4.2.1
pypng==0.0.18
PyQRCode==1.2.1
PyYAML==3.12
requests==2.18.1
urllib3==1.21.1
wxpy==0.3.9.8
